+++ CHANGES.rst
@@ -11,6 +11,8 @@ Changelog of fews-3di
 
 - Added evaporation upload.
 
+- Simulation is actually being run now.
+
 
 0.1 (2020-04-09)
 ----------------